requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
bcrypt==4.1.3
PyJWT==2.9.0
Flask-Limiter==3.8.0
python-dotenv==1.0.1